requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
base58check==1.0.2
bech32==1.2.0
black==22.6.0
blake256==0.1.1
cachetools==4.2.2
cbor==1.0.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
click==8.1.3
coinaddrvalidator==1.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Django==4.1.3
djangochannelsrestframework==1.1.0
djangorestframework==3.13.1
groestlcoin-hash2==1.1.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
install==1.3.5
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
Pyrogram==2.0.30
pysha3==1.0.2
PySocks==1.7.1
python-decouple==3.6
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rsa==4.9
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
supervisor==4.2.4
Telethon==1.24.0
TgCrypto==1.2.3
tomli==2.0.1
tornado==6.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
whitenoise==6.2.0
zope.interface==5.4.0
